move converter into 'test_features' which is about to move and rename, tests pass locally
