+++ zingolib/src/utils/conversion.rs
@@ -40,43 +40,3 @@ pub(crate) fn address_from_str(
 pub(crate) fn zatoshis_from_u64(amount: u64) -> Result<NonNegativeAmount, ConversionError> {
     NonNegativeAmount::from_u64(amount).map_err(|_e| ConversionError::OutsideValidRange)
 }
-
-/// Conversions for use in testing only
-#[cfg(test)]
-pub mod testing {
-    use zingoconfig::ChainType;
-
-    use crate::data::receivers::Receivers;
-
-    use super::{address_from_str, zatoshis_from_u64};
-
-    /// Converts primitive rust types to zcash receiver types for constructing transaction requests.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the address, amount or memo conversion fails.
-    pub fn receivers_from_send_inputs(
-        address_amount_memo_tuples: Vec<(&str, u64, Option<&str>)>,
-        chain: &ChainType,
-    ) -> Receivers {
-        address_amount_memo_tuples
-            .into_iter()
-            .map(|(address, amount, memo)| {
-                let recipient_address =
-                    address_from_str(address, chain).expect("should be a valid address");
-                let amount = zatoshis_from_u64(amount)
-                    .expect("should be inside the range of valid zatoshis");
-                let memo = memo.map(|memo| {
-                    crate::wallet::utils::interpret_memo_string(memo.to_string())
-                        .expect("should be able to interpret memo")
-                });
-
-                crate::data::receivers::Receiver {
-                    recipient_address,
-                    amount,
-                    memo,
-                }
-            })
-            .collect()
-    }
-}
